Move Spinner component to common directory
