+++ assets/js/common/Spinner/index.js
@@ -1,0 +1,3 @@
+import Spinner from './Spinner';
+
+export default Spinner;
