+++ assets/js/components/CleanUpButton/CleanUpButton.jsx
@@ -4,7 +4,7 @@ import { EOS_CLEANING_SERVICES } from 'eos-icons-react';
 import classNames from 'classnames';
 
 import Button from '@common/Button';
-import Spinner from '@components/Spinner';
+import Spinner from '@common/Spinner';
 
 function CleanUpButton({
   className,
